requirements.txt
antiorm==1.2.1
blinker==1.7.0
click==8.1.7
db==0.1.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pytesseract==0.3.10
regex==2023.12.25
SQLAlchemy==2.0.29
tqdm==4.66.2
typing_extensions==4.11.0
Werkzeug==3.0.2
